DDF-3499 Adding unit tests  - Removing unneeded setter
