+++ catalog/validator/catalog-validator-metacardwkt/pom.xml
@@ -44,6 +44,11 @@
             <artifactId>catalog-validator-wkt</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>com.vividsolutions</groupId>
+            <artifactId>jts</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
